+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Telegram-bot (KtGram) Changelog
 
-#### [Unreleased]
+## 7.6.0
 
+* Supported Telegram API [8.1](https://core.telegram.org/bots/api-changelog#december-4-2024).
 * Fixed `InlineQueryResult` cached and usual results serde clashing issue.
 * Fixed `savePreparedInlineMessage` options serde issue.
 * Added option to disable automatically cleaning class data through plugin's `autoCleanClassData` parameter.